Remove unused abstractions from the Java codegen
changelog_begin
changelog_end
