+++ language-support/java/codegen/src/test/scala/com/digitalasset/daml/lf/codegen/backend/java/JavaCodeGenTests.scala
@@ -1,42 +1,28 @@
 // Copyright (c) 2022 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.lf.codegen
+package com.daml.lf.codegen.backend.java
 
 import java.nio.file.Path
 
 import com.daml.bazeltools.BazelRunfiles
 import com.daml.lf.archive.DarReader
 import com.daml.lf.data.ImmArray.ImmArraySeq
-import com.daml.lf.data.Ref.{
-  DottedName,
-  ModuleName,
-  PackageId,
-  PackageName,
-  PackageVersion,
-  QualifiedName,
-}
-import com.daml.lf.iface.{DefDataType, Interface, InterfaceType, PackageMetadata, Record}
-import com.daml.lf.codegen.backend.java.JavaBackend
+import com.daml.lf.data.Ref._
+import com.daml.lf.iface._
 import com.daml.lf.codegen.conf.PackageReference
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.flatspec.AnyFlatSpec
 
-final class CodeGenTests extends AnyFlatSpec with Matchers {
+final class JavaCodeGenTests extends AnyFlatSpec with Matchers {
 
-  import CodeGenTests._
-
-  behavior of "backend"
-
-  it should "always use JavaBackend, which is currently hardcoded" in {
-    CodeGenRunner.backend should be theSameInstanceAs JavaBackend
-  }
+  import JavaCodeGenTests._
 
   behavior of "collectDamlLfInterfaces"
 
   it should "read interfaces from a single DAR file without a prefix" in {
 
-    val (interfaces, pkgPrefixes) = CodeGen.collectDamlLfInterfaces(Map(testDar -> None))
+    val (interfaces, pkgPrefixes) = JavaCodeGen.collectDamlLfInterfaces(Map(testDar -> None))
 
     assert(interfaces.length == 25)
     assert(pkgPrefixes == Map.empty)
@@ -44,7 +30,8 @@ final class CodeGenTests extends AnyFlatSpec with Matchers {
 
   it should "read interfaces from a single DAR file with a prefix" in {
 
-    val (interfaces, pkgPrefixes) = CodeGen.collectDamlLfInterfaces(Map(testDar -> Some("PREFIX")))
+    val (interfaces, pkgPrefixes) =
+      JavaCodeGen.collectDamlLfInterfaces(Map(testDar -> Some("PREFIX")))
 
     assert(interfaces.map(_.packageId).length == dar.all.length)
     assert(pkgPrefixes.size == dar.all.length)
@@ -55,7 +42,7 @@ final class CodeGenTests extends AnyFlatSpec with Matchers {
 
   it should "succeed if there are no collisions" in {
     assert(
-      CodeGen.detectModuleCollisions(
+      JavaCodeGen.detectModuleCollisions(
         Map.empty,
         Seq(interface("pkg1", "A", "A.B"), interface("pkg2", "B", "A.B.C")),
       ) === (())
@@ -64,7 +51,7 @@ final class CodeGenTests extends AnyFlatSpec with Matchers {
 
   it should "fail if there is a collision" in {
     assertThrows[IllegalArgumentException] {
-      CodeGen.detectModuleCollisions(
+      JavaCodeGen.detectModuleCollisions(
         Map.empty,
         Seq(interface("pkg1", "A"), interface("pkg2", "A")),
       )
@@ -73,7 +60,7 @@ final class CodeGenTests extends AnyFlatSpec with Matchers {
 
   it should "fail if there is a collision caused by prefixing" in {
     assertThrows[IllegalArgumentException] {
-      CodeGen.detectModuleCollisions(
+      JavaCodeGen.detectModuleCollisions(
         Map(PackageId.assertFromString("pkg2") -> "A"),
         Seq(interface("pkg1", "A.B"), interface("pkg2", "B")),
       )
@@ -82,7 +69,7 @@ final class CodeGenTests extends AnyFlatSpec with Matchers {
 
   it should "succeed if collision is resolved by prefixing" in {
     assert(
-      CodeGen.detectModuleCollisions(
+      JavaCodeGen.detectModuleCollisions(
         Map(PackageId.assertFromString("pkg2") -> "Pkg2"),
         Seq(interface("pkg1", "A"), interface("pkg2", "A")),
       ) === (())
@@ -107,7 +94,7 @@ final class CodeGenTests extends AnyFlatSpec with Matchers {
     val interface2 = interface(pkg2, Some(PackageMetadata(name2, version)))
     val interface3 = interface(pkg3, Some(PackageMetadata(name3, version)))
     assert(
-      CodeGen.resolvePackagePrefixes(
+      JavaCodeGen.resolvePackagePrefixes(
         pkgPrefixes,
         modulePrefixes,
         Seq(interface1, interface2, interface3),
@@ -121,12 +108,12 @@ final class CodeGenTests extends AnyFlatSpec with Matchers {
     val modulePrefixes =
       Map[PackageReference, String](PackageReference.NameVersion(name2, version) -> "A.B")
     assertThrows[IllegalArgumentException] {
-      CodeGen.resolvePackagePrefixes(Map.empty, modulePrefixes, Seq.empty)
+      JavaCodeGen.resolvePackagePrefixes(Map.empty, modulePrefixes, Seq.empty)
     }
   }
 }
 
-object CodeGenTests {
+object JavaCodeGenTests {
 
   private[this] val testDarPath = "language-support/java/codegen/test-daml.dar"
   private val testDar = Path.of(BazelRunfiles.rlocation(testDarPath))
